skills/requirements.txt
videodb
python-dotenv
pyyaml
